refactor(color-picker): splte directive into separate files, changed some names
