+++ src/lib/color-picker/color-picker-origin.directive.ts
@@ -2,7 +2,7 @@ import {
   AfterViewInit,
   ChangeDetectorRef,
   Directive,
-  ElementRef,
+  ElementRef, EventEmitter,
   forwardRef,
   Inject,
   Input,
@@ -14,56 +14,6 @@ import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { MccColorPickerComponent } from './color-picker.component';
 import { EMPTY_COLOR, ENABLE_ALPHA_SELECTOR, MccColorPickerOption, parseColorString, toHex, toRgba } from './color-picker';
 import { BehaviorSubject, Subscription } from 'rxjs';
-
-/**
- * This directive change the background of the button
- */
-@Directive({
-  selector: '[mccColorPickerOption], [mcc-color-picker-option]',
-  exportAs: 'mccColorPickerOption',
-})
-export class MccColorPickerOptionDirective implements AfterViewInit {
-  /**
-   * Receive the color
-   */
-  @Input('mccColorPickerOption')
-  get color(): MccColorPickerOption {
-    return this._color;
-  }
-  set color(value: MccColorPickerOption) {
-    this._color = value;
-  }
-  private _color: MccColorPickerOption;
-
-  constructor(
-    private elementRef: ElementRef,
-    private render: Renderer2,
-    @Inject(EMPTY_COLOR) private emptyColor: string
-  ) {
-    this._color = emptyColor;
-  }
-
-  ngAfterViewInit() {
-    if (this.color) {
-      let color: string;
-      if (typeof this.color === 'string') {
-        color = this.color;
-      } else {
-        color = this.color.value;
-        this.render.setAttribute(this.elementRef.nativeElement, 'aria-label', this.color.text);
-      }
-
-      if (parseColorString(color)) {
-        // apply the color
-        this.render.setStyle(
-          this.elementRef.nativeElement,
-          'backgroundColor',
-          color
-        );
-      }
-    }
-  }
-}
 
 /**
  * Directive applied to an element to make it usable as an origin for an ColorPicker.
@@ -83,7 +33,7 @@ export class MccColorPickerOriginDirective implements ControlValueAccessor {
   /**
    * Emit changes from the origin
    */
-  @Output() change: BehaviorSubject<string> = new BehaviorSubject<string>('');
+  @Output() change: EventEmitter<string> = new EventEmitter<string>();
 
   /**
    * Propagate changes to angular
